client: add types.ts into Fetcher dir
